+++ .prettierrc.js
@@ -1,7 +1,4 @@
 module.exports = {
-    semi: true,
-    trailingComma: "all",
-    singleQuote: true,
-    printWidth: 120,
+    ...require('@linzjs/style/.prettierrc.js'),
     tabWidth: 4
 };